Changed return type on getter from int to double.

+++ webapp/src/main/java/com/Data4Design/results/NumberResult.java
@@ -12,7 +12,7 @@ public class NumberResult extends Result{
 		this.data = (double)input;
 	}
 	
-	public int fetchData() {
+	public double fetchData() {
 		return data;
 	}
